+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Blight potion configs (found in dimension config files)
+### Fixed
+- Blight fires
+
 ## [2.0.2] - 2019-03-12
 ### Added
 - `sh_summon` command - Summons entities with a given difficulty and blight status, along with everything the vanilla summon command offers.